Add files via upload
THE DATA ENCRYPTION STANDARD (DES) algorithm is a symmetric-key block cipher that converts plaintext into ciphertext. It was developed in the early 1970s by an IBM team and adopted by the National Institute of Standards and Technology (NIST). The DES algorithm uses 64-bit plaintext blocks and 48-bit keys. The key length is 56 bits, but every eighth bit is used for parity checking and is ignored. The same algorithm and key are used for both encryption and decryption.